# swagger.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: My blogAPI
  description: This is a very simple and minimalistic blog api for posting articles and get comment from users
servers:
- url: http://localhost:4000/api/v1
  description: dev serve
...
paths:
...
  /comments/{id}:
    get:
      security:
        - bearerauth: []
      tags: ['Comment']
      description: get a particular comment 
      parameters:
        - $ref: '#/components/parameters/idPathParam'
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: {type: string, examples: ['successful']}
                  data: 
                    allOf:
                      - type: object
                        properties:
                          id : {type: string, examples: ['6585xe74seq454ax26975']}
                          body : {type: string, examples: ['this is a great discussion']}
                          author: 
                            $ref: '#/components/schemas/AuthorDTO'
                          article: {type: string, examples: ['6585xe74seq454ax26975']}
                      -  $ref: '#/components/schemas/Timestamp'
                  links:
                    type: object
                    properties:
                      self: {type: string, examples: ['/comments/6585xe74seq454ax26975']}
                      author: {type: string, examples: ['/profile/6585xe74seq454ax26975']}
                      article: {type: string, examples: ['/articles/6585xe74seq454ax26975']}
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
    patch:
      security:
        - bearerauth: []
      tags: ['Comment']
      description: update a comments
      parameters:
        - $ref: '#/components/parameters/idPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                body: {type: string, format: text, examples: ['this was not the case']}
                status: {type: string, enum: ['public', 'private'], examples: ['private']}
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: {type: string, examples: ['successful']}
                  data: 
                    allOf:
                      - type: object
                        properties:
                          id : {type: string, examples: ['6585xe74seq454ax26975']}
                          body : {type: string, examples: ['this is a great discussion']}
                          author: 
                            $ref: '#/components/schemas/AuthorDTO'
                          article: {type: string, examples: ['6585xe74seq454ax26975']}
                      -  $ref: '#/components/schemas/Timestamp'
                  links:
                    type: object
                    properties:
                      self: {type: string, examples: ['/comments/6585xe74seq454ax26975']}
                      author: {type: string, examples: ['/profile/6585xe74seq454ax26975']}
                      article: {type: string, examples: ['/articles/6585xe74seq454ax26975']}
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
    delete:
      security:
        - bearerauth: []
      tags: ['Comment']
      description: delete a particular comment
      parameters:
        - $ref: '#/components/parameters/idPathParam'
      responses: 
        '204':
          description: Article deleted successfully
        '401':
          $ref: '#/components/responses/401'
        '404':
          $ref: '#/components/responses/404'      
...
components:
  securitySchemes:
    bearerauth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Timestamp:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AuthorDTO:
      type: object
      required:
      - id
      - name
      properties:
        id:
          type: string
          examples:
          - '6585xe74seq454ax26975'
        name:
          type: string
          examples:
          - Raybrin Max
  responses:
    "400":
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                examples:
                - 400
              error:
                type: string
                examples:
                - Bad request
              data:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      examples:
                      - email
                    message:
                      type: string
                      examples:
                      - invalid email
    "401":
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                examples:
                - 401
              error:
                type: string
                examples:
                - Unauthorized request
              message:
                type: string
                examples:
                - You don't have the permission
    "404":
      description: Resource Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                examples:
                - 404
              error:
                type: string
                examples:
                - Not Found
              message:
                type: string
                examples:
                - Request resource not found
...
  parameters:
...
    idPathParam:
      name: id
      required: true
      in: path
      description: Id
      schema:
        type: string
        examples:
        - 125455